feat(evm): avoid overflow on expiry if trustingperiod is set to max(uint64)

+++ evm/contracts/clients/CometblsClient.sol
@@ -68,7 +68,7 @@ library CometblsClientLib {
         uint64 trustingPeriod,
         uint64 currentTime
     ) internal pure returns (bool) {
-        return currentTime > (headerTime + trustingPeriod);
+        return uint256(currentTime) > (uint256(headerTime) + uint256(trustingPeriod));
     }
 
     function encodeMemory(
